few windows updates
git-svn-id: https://svn.resiprocate.org/rep/resiprocate/main@306 ddefafc4-47db-0310-ae44-fa13212b10f2

+++ sip/resiprocate/Transport.cxx
@@ -78,7 +78,7 @@ Transport::buildFdSet( fd_set* fdSet, int* fdSetSize )
    assert( fdSetSize );
 	
    FD_SET(mFd,fdSet);
-   if ( mFd <= *fdSetSize )
+   if ( mFd <= (unsigned int)(*fdSetSize) )
    {
       *fdSetSize = mFd+1;
    }
